race.describe {race}	R Documentation

Describe a candidate

Description

This function may be provided by the user for giving a description of a candidate. It's definition has to be given in the same file in which the functions race.wrapper and race.info are defined. The name of such file has to be passed as first argument to the function race.

Usage

race.describe(candidate,data)

Arguments

candidate 	 The candidate for which a description is to be returned.

data 	 It is the object of type list (possibly empty) returned by race.init, if the latter is defined by the user.

Value

The function race.describe should return an object describing the selected candidate. Such object will be printed by race through the function print.

Author(s)

Mauro Birattari

See Also

race, race.init

Examples

# Please have a look at the function `race.describe'
# defined in the file `example-wrapper.R':
local({
  source(file.path(system.file(package="race"),
                           "examples","example-wrapper.R"),local=TRUE);
  print(race.describe)})

[Package race version 0.1.55 Index]