race.info {race}	R Documentation

Provide information on the race

Description

This function is to be provided by the user. It's definition has to be given (together with the one of race.wrapper) in a file, and the name of such file has to be passed as first argument to the function race.

Usage

race.info(data)

Arguments

data 	 It is the object of type list (possibly empty) returned by race.init, if the latter is defined by the user.

Value

The function race.info is expected to return a list including the following components:

race.name 	 The name of the race for documentation purposes.

no.candidates 	 The number of candidates in the race.

no.tasks 	 Number of tasks available for testing.

no.subtasks 	 Each task might consists of no.subtasks subtasks. If the element no.subtasks is not included in the list, it is assumed that each task is indeed atomic, that is, no.subtasks=1. no.subtasks may also be a vector of length no.tasks. In this case, the i-th task consists of no.subtasks[i] subtasks.

extra 	 A character string providing extra information on the race for documentation purposes. It can be a long string and the user is not required to introduce newline characters: it will be automatically formatted for pretty-printing.

Author(s)

Mauro Birattari

See Also

race, race.init

Examples

# Please have a look at the function `race.info'
# defined in the file `example-wrapper.R':
local({
  source(file.path(system.file(package="race"),
                           "examples","example-wrapper.R"),local=TRUE);
  print(race.info)})

[Package race version 0.1.55 Index]