race.init {race}	R Documentation

Initialization function

Description

This function may be provided by the user for initializing the computation of the slave processes. It's definition has to be given in the same file in which the functions race.wrapper and race.info are defined. The name of such file has to be passed as first argument to the function race.

Usage

race.init()

Arguments

The function race.init has to be called with no arguments.

Details

This function should be used for initializing the computation on each slave, e.g. loading libraries or data needed by race.wrapper, race.info, and/or race.describe. The output of race.init will be passed to these functions.

Value

The function race.init is expected to return an object of mode list.

Author(s)

Mauro Birattari

See Also

race, race.wrapper, race.info, race.describe

Examples

# Please have a look at the function `race.init'
# defined in the file `example-wrapper.R':
local({
  source(file.path(system.file(package="race"),
                           "examples","example-wrapper.R"),local=TRUE);
  print(race.init)})

[Package race version 0.1.55 Index]